added dependencies for data layer
