+++ app/src/main/java/luci/sixsixsix/powerampache2/common/Extensions.kt
@@ -68,32 +68,9 @@ import java.lang.ref.WeakReference
 import java.security.MessageDigest
 import kotlin.math.absoluteValue
 
-
-typealias WeakContext = WeakReference<Context>
-
-fun String.md5(): String {
-    return hashString(this, "MD5")
-}
-
-fun String.sha256(): String {
-    return hashString(this, "SHA-256")
-}
-
-fun String.isIpAddress(): Boolean =
-    //InetAddresses.isNumericAddress(this) // requires min Q
-    Patterns.IP_ADDRESS.matcher(this).matches()
-    //("(\\d{1,2}|(0|1)\\" + "d{2}|2[0-4]\\d|25[0-5])").toRegex().matches(this)
-
 val Int.dpTextUnit: TextUnit
     @Composable
     get() = with(LocalDensity.current) { this@dpTextUnit.dp.toSp() }
-
-private fun hashString(input: String, algorithm: String): String {
-    return MessageDigest
-        .getInstance(algorithm)
-        .digest(input.toByteArray())
-        .fold("") { str, it -> str + "%02x".format(it) }
-}
 
 fun Context.shareLink(link: String) =
     startActivity(
@@ -149,81 +126,6 @@ fun Context.exportSong(song: Song, offlineUri: String) {
 }
 
 
-/**
- * TODO DEBUG
- * this function uses reflection to quickly turn any of the music object into a String to quickly
- * print and visualize data
- */
-fun Any.toDebugString(
-    excludeErrorValues: Boolean = false,
-    excludeLyrics: Boolean = false,
-    excludeLists: Boolean = false,
-    separator: String = "\n"
-): String {
-    val obj = this
-    val sb = StringBuilder()
-    for (field in obj.javaClass.declaredFields) {
-        field.isAccessible = true
-
-        field.get(obj)?.let {
-            if(
-                !field.name.lowercase().contains("url") &&
-                !field.name.lowercase().contains("token") &&
-                !field.name.lowercase().contains("artist") &&
-                !field.name.lowercase().contains("CREATOR") &&
-                !field.name.lowercase().contains("\$stable") &&
-                "$it".isNotBlank() &&
-                "$it" != "0" &&
-                !"$it".contains("CREATOR") &&
-                !"$it".contains("\$stable") &&
-                "$it" != "[]" &&
-                (!excludeLyrics || !field.name.lowercase().contains("lyrics")) &&
-                (!excludeLists || it !is List<*>)
-            ) {
-                if (it is List<*>) {
-                    if (field.name != "genre") {
-                        sb.append(field.name)
-                            .append(": ")
-                    } else {
-                        sb.append(" | ")
-                    }
-
-                    it.forEach { listElem ->
-                        listElem?.let {
-                            if (listElem is MusicAttribute) {
-                                sb.append(listElem.name)
-                                sb.append(" | ")
-                            }
-                        }
-                    }
-                    sb.append(separator)
-                } else if (it is MusicAttribute) {
-                    sb.append(field.name)
-                        .append(": ")
-                        .append("${it.name}")
-                        .append(separator)
-                } else {
-                    val valueStr = "${field.get(obj)}"
-                    if (!excludeErrorValues ||
-                            (!valueStr.startsWith(ERROR_STRING) &&
-                             !valueStr.startsWith(ERROR_TITLE) &&
-                             !valueStr.startsWith(ERROR_INT.toString()) &&
-                             !valueStr.startsWith(ERROR_FLOAT.toString())
-                            )
-                    ) {
-                        sb.append(field.name)
-                            .append(": ")
-                            .append(valueStr)
-                            .append(separator)
-                    }
-                }
-            }
-        }
-    }
-    // remove variables that are auto generate by the parcelable
-    return sb.toString().split("CREATOR")[0]
-}
-
 fun getVersionInfoString(context: Context) = try {
     val pInfo: PackageInfo =
         context.packageManager.getPackageInfo(context.packageName, 0)
@@ -233,70 +135,10 @@ fun getVersionInfoString(context: Context) = try {
     ""
 } + " - DB: ${Constants.DATABASE_VERSION}"
 
-fun Any.toDebugMap() = LinkedHashMap<String, String>().also { map ->
-    val separator = "---666---"
-    val debugStr = toDebugString(excludeLists = true, excludeLyrics = true, excludeErrorValues = true, separator = separator)
-    debugStr.split(separator).forEach {
-        if (!it.startsWith("Companion")) {
-            val keyValue = it.split(": ")
-            if (keyValue.size == 2) {
-                map[keyValue[0]] = keyValue[1]
-            }
-        }
-    }
-}
-
 @Composable
 @ReadOnlyComposable
 fun fontDimensionResource(@DimenRes id: Int) = dimensionResource(id = id).value.sp
 
-fun processFlag(flag: Any?): Int = flag?.let {
-    if (it is Boolean) {
-        if (it == true) 1 else 0
-    } else if (it is Int) it
-    else 0
-} ?: 0
-
-/**
- * different versions of ampache seem to return different types that are making gson crash
- */
-fun processNumberToInt(num: Any?): Int = num?.let {
-    try {
-        when(it) {
-            is Double -> it.toInt()
-            is Float -> it.toInt()
-            is Int -> it.toInt()
-            is String -> it.toInt()
-            else -> ERROR_INT
-        }
-    } catch (e: Exception) { ERROR_INT }
-} ?: ERROR_INT
-
-/**
- * different versions of ampache seem to return different types that are making gson crash
- */
-fun processNumberToFloat(num: Any?): Float = num?.let {
-    try {
-        when(it) {
-            is Double -> it.toFloat()
-            is Float -> it.toFloat()
-            is Int -> it.toFloat()
-            is String -> it.toFloat()
-            else -> ERROR_FLOAT
-        }
-    } catch (e: Exception) {
-        ERROR_FLOAT
-    }
-} ?: ERROR_FLOAT
-
-/**
- * if hasArt flag is present (not null), check if there's any art with it.
- */
-fun processArtUrl(hasArt: Any?, artUrl: String?) = hasArt?.let { hasArtAny ->
-    if (artUrl != null && processFlag(hasArtAny) == 1) { artUrl } else { "" }
-} ?: run {
-    artUrl ?: ""
-}
 
 @ColorInt
 fun String.toHslColor(saturation: Float = 0.5f, lightness: Float = 0.4f): Int {
